# Add comprehensive README test concept & update GitHub Issue #3 critical security test performance problems