--- a/.claude/commands/execute-tasks.md
+++ b/.claude/commands/execute-tasks.md
@@ -156,62 +156,39 @@ Regular checks during execution:
 - Validate current state before continuing
 - Handle partially completed tasks
 
-## Final Validation & PR Creation
+## Task Completion Notification
 
-### Complete Feature Validation
-Before creating PR, ensure:
-- All tasks marked as completed
-- Full test suite passes (100% success rate)
-- Code coverage meets requirements
-- All quality gates pass
-- Documentation is updated
+### Complete Task Execution
+When all tasks completed, notify user that execution phase is finished:
 
-### Pull Request Creation
-When all tasks completed:
-
-**PR Title Format:**
+```bash
+echo "🎉 All tasks completed successfully!"
+echo "📊 Final Statistics:"
+echo "   - Tasks completed: {total_completed}"
+echo "   - Test coverage: {final_coverage}%"
+echo "   - Commits created: {commit_count}"
+echo ""
+echo "📋 Next Steps:"
+echo "   1. Run final validation: pytest --tb=short"
+echo "   2. Complete feature: /project:complete-feature {SUFFIX}"
+echo ""
+echo "🔧 Optional Manual Testing:"
+echo "   - python test_mcp_tool_call.py"
+echo "   - mcp-inspector mcp-inspector-config.json"
 ```
-feat({scope}): {feature_summary}
-```
 
-**PR Description Template:**
+### Progress File Finalization
+Mark the progress file as completed:
 ```markdown
-## Feature Summary
-{Brief description of implemented feature}
+## Status
+- **Current Phase**: COMPLETED
+- **Current Task**: ALL TASKS COMPLETED
+- **Overall Progress**: 100%
+- **Completed**: {completion_timestamp}
 
-## Implementation Details
-- **Source**: {original_issue_or_file}
-- **Tasks Completed**: {total_task_count}
-- **Test Coverage**: {final_coverage}%
-- **Commits**: {commit_count}
-
-## Testing
-- [x] Unit tests passing
-- [x] Integration tests passing
-- [x] System tests passing
-- [x] Performance tests passing
-- [x] Security tests passing
-
-## Quality Gates
-- [x] Linting passed
-- [x] Type checking passed
-- [x] Code coverage >= 80%
-- [x] Documentation updated
-
-## Changes Made
-{Auto-generated summary of key changes}
-
-## Testing Instructions
-{How to test the feature}
-
-Closes #{issue_number}
+## Final Summary
+All planned tasks successfully completed. Ready for feature completion workflow.
 ```
-
-**PR Labels & Metadata:**
-- Automatic labels based on feature type
-- Link to original issue
-- Assign to appropriate milestone
-- Add relevant project boards
 
 ## Error Handling & Recovery