--- a/start_server.sh
+++ b/start_server.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+cd /Users/florianwegener/Projects/crawl4ai-mcp-server
+exec uv run python server.py