… .claude/commands/complete-feature.md
… .planning/PROGRESS_consolidate-testing-knowledge.md
… __pycache__/server.cpython-313.pyc
… pyproject.toml
… tests/__pycache__/test_mcp_protocol.cpython-313-pytest-8.4.1.pyc
… tests/__pycache__/test_models.cpython-313-pytest-8.4.1.pyc
… tests/__pycache__/test_project_structure.cpython-313-pytest-8.4.1.pyc
… tests/__pycache__/test_server.cpython-313-pytest-8.4.1.pyc
… tests/__pycache__/test_testing_framework.cpython-313-pytest-8.4.1.pyc
… tests/__pycache__/test_web_extract.cpython-313-pytest-8.4.1.pyc
… tools/__pycache__/__init__.cpython-313.pyc
… tools/__pycache__/web_extract.cpython-313.pyc